+++ electron/app/createWindow.ts
@@ -5,18 +5,12 @@ import {machineIdSync} from 'node-machine-id';
 import Nucleus from 'nucleus-nodejs';
 import * as path from 'path';
 
-import {
-  APP_MIN_HEIGHT,
-  APP_MIN_WIDTH,
-  DEFAULT_PLUGINS,
-  DEFAULT_TEMPLATES_PLUGIN_URL,
-  DEPENDENCIES_HELP_URL,
-  NEW_VERSION_CHECK_INTERVAL,
-} from '@constants/constants';
-
 import {disableSegment, enableSegment, getSegmentClient} from '@utils/segment';
 
+import {APP_MIN_HEIGHT, APP_MIN_WIDTH, NEW_VERSION_CHECK_INTERVAL} from '@monokle-desktop/shared/constants/app';
+import {DEFAULT_PLUGINS} from '@monokle-desktop/shared/constants/plugin';
 import {DISABLED_TELEMETRY} from '@monokle-desktop/shared/constants/telemetry';
+import {DEFAULT_TEMPLATES_PLUGIN_URL, DEPENDENCIES_HELP_URL} from '@monokle-desktop/shared/constants/urls';
 import {AlertEnum, NewVersionCode, StartupFlags} from '@monokle-desktop/shared/models';
 import type {AlertType} from '@monokle-desktop/shared/models';
 import utilsElectronStore from '@monokle-desktop/shared/utils/electronStore';
