+++ shared/src/constants/app.ts
@@ -1,0 +1,3 @@
+export const APP_MIN_WIDTH = 800;
+export const APP_MIN_HEIGHT = 600;
+export const NEW_VERSION_CHECK_INTERVAL = 24 * 60 * 60 * 1000;
