+++ shared/src/constants/index.ts
@@ -1,5 +1,8 @@
+export * from './app';
 export * from './electronStore';
 export * from './hotkeys';
 export * from './ipcEvents';
 export * from './k8s';
+export * from './plugin';
 export * from './telemetry';
+export * from './urls';
