+++ shared/src/constants/plugin.ts
@@ -1,0 +1,9 @@
+import {DEFAULT_TEMPLATES_PLUGIN_URL} from './urls';
+
+export const DEFAULT_PLUGINS = [
+  {
+    owner: 'kubeshop',
+    name: 'monokle-default-templates-plugin',
+    url: DEFAULT_TEMPLATES_PLUGIN_URL,
+  },
+];
