+++ src/components/organisms/PluginManager/PluginManager.tsx
@@ -6,11 +6,13 @@ import {Button, Skeleton, Tooltip, Typography} from 'antd';
 
 import {PlusOutlined, ReloadOutlined} from '@ant-design/icons';
 
-import {DEFAULT_TEMPLATES_PLUGIN_URL, PLUGINS_HELP_URL, TOOLTIP_DELAY} from '@constants/constants';
+import {PLUGINS_HELP_URL, TOOLTIP_DELAY} from '@constants/constants';
 import {PluginManagerDrawerReloadTooltip} from '@constants/tooltips';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {checkForExtensionsUpdates} from '@redux/services/extension';
+
+import {DEFAULT_TEMPLATES_PLUGIN_URL} from '@monokle-desktop/shared/constants/urls';
 
 import PluginInformation from './PluginInformation';
 import PluginInstallModal from './PluginInstallModal';
